requirements.txt
streamlit
pandas
gspread
google-auth
openpyxl
matplotlib
streamlit-authenticator
streamlit-authenticator==0.2.2
gspread
google-auth
plotly
xlsxwriter

